SERVER-8412 clean up intmax_t
Conflicts:
	src/mongo/db/dbcommands.cpp
	src/mongo/db/repair_database.cpp
	src/mongo/db/repl/oplog.cpp

+++ src/mongo/db/database.cpp
@@ -77,7 +77,7 @@ namespace mongo {
         }
     }
 
-    void Database::diskSize(size_t &uncompressedSize, size_t &compressedSize) {
+    void Database::diskSize(intmax_t &uncompressedSize, intmax_t &compressedSize) {
         list<string> colls;
         _collectionMap.getNamespaces(colls);
         CollectionData::Stats dbstats;
